Update interface.ts
The import statement for ChatCompletionRequestMessage is unnecessary if it's not used within this file.
The IConfig interface has a required property openai_api_key, but api is optional. It might be preferable to make openai_api_key optional as well for more flexibility.
The User interface defines chatMessage as an array of ChatCompletionRequestMessage, but it seems like ChatCompletionRequestMessage is an import rather than a defined interface or type.
There are no comments to explain the purpose of each interface or any other parts of the code.